move Java known issues into a separate file to use a PR file filter
