+++ azure-pipelines/java-preview-compile-tests-known-issues.yml
@@ -12,6 +12,7 @@ pr:
       - Java*
       - msgraph-sdk-raptor-compiler-lib/MicrosoftGraphJavaCompiler.cs
       - TestsCommon/JavaTestRunner.cs
+      - TestsCommon/JavaKnownIssues.cs
 schedules:
   - cron: "0 3 * * *" # everyday at 3AM UTC (off hours for Redmond, Nairobi and Montréal)
     displayName: 'Daily Preview Java Snippet Compilation Tests - Known Issues'
